修改表接口

+++ database/migrations/2018_05_29_172540_create_consult_tels_table.php
@@ -15,7 +15,7 @@ class CreateConsultTelsTable extends Migration
     {
         Schema::create('consult_tels', function (Blueprint $table) {
             $table->increments('id');
-            $table->tinyInteger('user_id')->nullable()->comment('用户id');
+            $table->integer('user_id')->nullable()->comment('用户id');
             $table->string('source',32)->nullable()->comment('来源');
             $table->string('page_source',32)->nullable()->comment('页面来源');
             $table->timestamps();
